feat: Add web support with informative message and update dependencies
- Implemented web support in the app, displaying a message when accessed via a browser.
- Updated model manager to use new download method.
- Refactored library cubit to use new repository methods.
- Added new HTML and manifest files for web deployment.
- Included favicon and icons for web app.
- Adjusted main entry point to skip initialization on web.
- Updated various files to change permissions and modes.

+++ lib/features/setup/model_setup_cubit.dart
@@ -51,7 +51,7 @@ class ModelSetupCubit extends Cubit<ModelSetupState> {
     emit(state.copyWith(status: SetupStatus.downloading, progress: 0));
     try {
       await for (final pct in _modelManager.downloadModel()) {
-        emit(state.copyWith(progress: pct));
+        emit(state.copyWith(progress: pct.toDouble()));
       }
       emit(state.copyWith(status: SetupStatus.ready));
     } catch (e) {
